# Convert the `Dict`-implementation to use a `Map` internally