--- a/.prettierignore
+++ b/.prettierignore
@@ -6,6 +6,7 @@ external/bcmaps/
 external/builder/fixtures/
 external/builder/fixtures_babel/
 external/quickjs/
+test/stats/results/
 test/tmp/
 test/pdfs/
 web/locale/